Split AuraTimerOverlayFeature into TowerAuraTimerOverlayFeature and TowerAuraVignetteFeature
