+++ common/src/main/java/com/wynntils/features/combat/TowerAuraVignetteFeature.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.features.combat;
+
+import com.wynntils.core.components.Models;
+import com.wynntils.core.config.Category;
+import com.wynntils.core.config.Config;
+import com.wynntils.core.config.ConfigCategory;
+import com.wynntils.core.config.RegisterConfig;
+import com.wynntils.core.consumers.features.Feature;
+import com.wynntils.mc.event.RenderEvent;
+import com.wynntils.utils.MathUtils;
+import com.wynntils.utils.colors.CommonColors;
+import com.wynntils.utils.colors.CustomColor;
+import com.wynntils.utils.render.RenderUtils;
+import net.minecraftforge.eventbus.api.EventPriority;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+@ConfigCategory(Category.COMBAT)
+public class TowerAuraVignetteFeature extends Feature {
+    private static final float MAX_INTENSITY = 0.4f;
+
+    @RegisterConfig
+    public final Config<Boolean> vignetteOnAura = new Config<>(true);
+
+    @RegisterConfig
+    public final Config<CustomColor> vignetteColor = new Config<>(CommonColors.ORANGE);
+
+    @SubscribeEvent(priority = EventPriority.LOWEST)
+    public void onRenderGui(RenderEvent.Post event) {
+        if (!vignetteOnAura.get() || event.getType() != RenderEvent.ElementType.GUI) return;
+
+        long remainingTimeUntilAura = Models.TowerAuraTimer.getRemainingTimeUntilAura();
+        if (remainingTimeUntilAura <= 0) return;
+
+        RenderUtils.renderVignetteOverlay(
+                event.getPoseStack(),
+                vignetteColor.get(),
+                MathUtils.map(remainingTimeUntilAura, Models.TowerAuraTimer.getAuraLength(), 0, 0, MAX_INTENSITY));
+    }
+}
