+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -44,17 +44,6 @@
   "feature.wynntils.arrowShieldTrackerOverlay.overlay.arrowShieldTracker.textColor.name": "Text Color",
   "feature.wynntils.arrowShieldTrackerOverlay.overlay.arrowShieldTracker.textShadow.description": "What should the text shadow look like?",
   "feature.wynntils.arrowShieldTrackerOverlay.overlay.arrowShieldTracker.textShadow.name": "Text Shadow",
-  "feature.wynntils.auraTimerOverlay.description": "Adds a timer to see when the next guild war tower aura will hit.",
-  "feature.wynntils.auraTimerOverlay.name": "Aura Timers",
-  "feature.wynntils.auraTimerOverlay.overlay.auraTimer.name": "Aura Timer",
-  "feature.wynntils.auraTimerOverlay.overlay.auraTimer.textColor.description": "What color should the aura timer text be?",
-  "feature.wynntils.auraTimerOverlay.overlay.auraTimer.textColor.name": "Text Color",
-  "feature.wynntils.auraTimerOverlay.overlay.auraTimer.textShadow.description": "What should the text shadow look like?",
-  "feature.wynntils.auraTimerOverlay.overlay.auraTimer.textShadow.name": "Text Shadow",
-  "feature.wynntils.auraTimerOverlay.vignetteColor.description": "What color should the vignette overlay be?",
-  "feature.wynntils.auraTimerOverlay.vignetteColor.name": "Vignette Color",
-  "feature.wynntils.auraTimerOverlay.vignetteOnAura.description": "Should a vignette overlay be applied when an aura is active?",
-  "feature.wynntils.auraTimerOverlay.vignetteOnAura.name": "Vignette Overlay",
   "feature.wynntils.autoApplyResourcePack.description": "Adds the ability to automatically apply the Wynntils resource pack on startup.",
   "feature.wynntils.autoApplyResourcePack.name": "Auto Apply Resource Pack",
   "feature.wynntils.autoJoinParty.description": "Adds the ability to automatically join a party when a friend invites you.",
@@ -993,6 +982,19 @@
   "feature.wynntils.tooltipVanillaHide.hideAdvanced.description": "This will turn off advanced tooltips (F3+H) on Wynncraft",
   "feature.wynntils.tooltipVanillaHide.hideAdvanced.name": "Always Hide Advanced Tooltip",
   "feature.wynntils.tooltipVanillaHide.name": "Hide Irrelevant Tooltip Info",
+  "feature.wynntils.towerAuraTimerOverlay.description": "Adds a timer to see when the next guild war tower aura will hit.",
+  "feature.wynntils.towerAuraTimerOverlay.name": "Tower Aura Timer Overlay",
+  "feature.wynntils.towerAuraTimerOverlay.overlay.towerAuraTimer.name": "Aura Timer",
+  "feature.wynntils.towerAuraTimerOverlay.overlay.towerAuraTimer.textColor.description": "What color should the aura timer text be?",
+  "feature.wynntils.towerAuraTimerOverlay.overlay.towerAuraTimer.textColor.name": "Text Color",
+  "feature.wynntils.towerAuraTimerOverlay.overlay.towerAuraTimer.textShadow.description": "What should the text shadow look like?",
+  "feature.wynntils.towerAuraTimerOverlay.overlay.towerAuraTimer.textShadow.name": "Text Shadow",
+  "feature.wynntils.towerAuraVignette.description": "Adds a vignette when when a tower aura is active.",
+  "feature.wynntils.towerAuraVignette.name": "Tower Aura Vignette",
+  "feature.wynntils.towerAuraVignette.vignetteColor.description": "What color should the vignette overlay be?",
+  "feature.wynntils.towerAuraVignette.vignetteColor.name": "Vignette Color",
+  "feature.wynntils.towerAuraVignette.vignetteOnAura.description": "Should a vignette overlay be applied when an aura is active?",
+  "feature.wynntils.towerAuraVignette.vignetteOnAura.name": "Vignette Overlay",
   "feature.wynntils.tradeMarketAutoOpenChat.description": "Automatically open chat when using the trade market.",
   "feature.wynntils.tradeMarketAutoOpenChat.name": "Market Auto Open Chat",
   "feature.wynntils.tradeMarketBulkSell.bulkSell1Amount.description": "How many items should be sold when the bulk sell 1 button is pressed?",
